Move Unrestricted, Movable to their own module
Move Unrestricted, Consumable, Dupable, Movable to their own module so
that they can be used internally without importing all of Prelude.Linear
which often causes import cycles
